Fixed GA, implemented basic plotting
I've fixed the issues with GA generations, resulting in code that can take an initial population and (optionally) an original puzzle to match, and produces solutions in an attempt to create an acceptable sudoku solution, matching the original puzzle if one is given. I have also created the file mpl.py, which contains code for using matplotlib to plot data from running this GA code.

Additionally, both ga.py and mpl.py now have test case functions, which can be used to demonstrate the code and to test its capabilities. Note in particular that test_case_2() in ga.py highlights a major flaw of our current approach to GA: the mutation function does not introduce new cell values that were not present in the initial population. For this reason, I recommend we strongly consider incorporating other mutation functions as well, such as randomly altering a cell (possibly prioritizing empty cells) to a new value.

I also made a few changes to sudoku.py, but none that should affect functionality with previous code. The changes that I made to ga.py DO affect functionality, make sure that if you call the changed functions, that your arguments are declared in a way to work around the newly-added puzzle argument.

The functions I've added to sudoku.py are as follows:

- empty_count(self): returns the total number of empty cells remaining in this sudoku puzzle
- matches_original(self, original): returns True if self is a valid (possibly partial) solution to the original puzzle, and False otherwise (for example, if it overwrites a cell that was given a starting value in the original puzzle)
- print_grid(grid, grid_size): prints a grid directly, without needing to take a Sudoku object - convenient for debugging

Changes to ga.py that DO affect functionality:

- default_evaluation now takes a puzzle argument, which defaults to None
- iterate_one_generation, as well as all variations on run_ga, now take a puzzle argument if they did not already do so; if no such argument is provided, it defaults to None, and evaluation is handled as though there were no original puzzle to solve

The new content in mpl.py:

- run_ga_plot_data: a method that works much like run_ga_get_data, but this one will plot the data with matplotlib (remember that you need to call pyplot.show() to view this plot; multiple plots can be drawn before viewing the result, this way)
- the above method accepts a mode argument, which defines how data is to be gathered; this can be set to gather the average or best sum of wvalues in a generation, and can display data from every generation, only every generation matching a power of 10, or only the final generation

TODO:
- implement alternative GA methods
- ensure algorithm effectiveness working with initial puzzles
- ensure algorithm can introduce values not in initial population
-